requirements.txt
# ============================================================
# Omni-ST: Instruction-Driven Any-to-Any Multimodal Modeling
# for Spatial Transcriptomics — Requirements
# ============================================================

# Core Deep Learning
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Transformer Architectures
transformers>=4.36.0
timm>=0.9.12
einops>=0.7.0

# Graph Neural Networks
torch-geometric>=2.4.0
torch-scatter>=2.1.2
torch-sparse>=0.6.18

# Spatial Transcriptomics
scanpy>=1.9.6
anndata>=0.10.4
squidpy>=1.3.1

# Scientific Computing
numpy>=1.26.0
scipy>=1.11.4
scikit-learn>=1.3.2
pandas>=2.1.4

# Visualization
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.18.0
umap-learn>=0.5.5
streamlit>=1.29.0
dash>=2.14.2

# Experiment Tracking
wandb>=0.16.1
hydra-core>=1.3.2
omegaconf>=2.3.0

# Training Utilities
accelerate>=0.25.0
deepspeed>=0.12.6
peft>=0.7.1

# Image Processing
Pillow>=10.1.0
opencv-python>=4.8.1.78
albumentations>=1.3.1

# Bioinformatics
pydeseq2>=0.4.5
harmonypy>=0.0.9

# Notebook
jupyter>=1.0.0
ipykernel>=6.28.0
ipywidgets>=8.1.1

# Testing & Code Quality
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.12.0
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0

# Utilities
tqdm>=4.66.1
rich>=13.7.0
pydantic>=2.5.3
h5py>=3.10.0
zarr>=2.16.1
requests>=2.31.0
